modularise the eng module somewhat
