+++ ogma/Cargo.toml
@@ -2,6 +2,7 @@
 name = "ogma"
 version = "0.1.0"
 authors = ["kurt <kurtlawrence92@gmail.com>"]
+# TODO -- update to 2021
 edition = "2018"
 
 [dependencies]
