+++ ogma/src/eng/ty/mod.rs
@@ -1,0 +1,3 @@
+//! Typing annotation and inferencing.
+
+mod annotate;
